+++ src/main/java/org/expense_bot/service/IncomeService.java
@@ -2,12 +2,13 @@ package org.expense_bot.service;
 
 import org.expense_bot.model.Income;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
 public interface IncomeService {
 
-  Income save(Long userId, Double sum, LocalDateTime incomeDate);
+  Income save(Long userId, BigDecimal sum, LocalDateTime incomeDate);
 
   List<Income> getAllCurrentMonth(Long userId);
